feat: fast-forward to Flagrum 1.6.0
Flagrum was temporarily separated into a private repository around v1.4.3 to protect the Patreon benefits from being easily recompiled from the codebase. This commit is a squash merge of all changes since then up to v1.6.0 to return Flagrum to the public domain. This was made possible due to a large refactor of the codebase that allows the Patreon benefits to remain in a separate repository, while leaving the free edition in perfect working order without the need to have access to this private repository.