fix: accomodate release mod loader path
Mod loader DLL appears in a different location on release builds, likely due to releasing specifically for win-x64 so not needing the runtimes folder. This commit ensures that the root application directory is now checked for the mod loader first, and will fall back to the runtimes directory if this does not exist (such as when running locally on Any CPU).