fix: resolve demo file index crashes
Fixes a bug caused by indexed loose files not being converted to lowercase by the file indexer.

Ensures AssetIds match between Flagrum and the file index for demo and PS4 versions of the game by recomputing the hash in EbonyArchiveReader rather than using the precomputed hash. This is a band-aid fix, but a real solution would take a lot of time and effort.

Ensures indexed loose files don't cause issues by now limiting this to a whitelist of allowed file extensions. Will require users that indexed the demos in the past to reindex those profiles to resolve the problem, but as this is a rare use-case, it's a non-issue.

Refs: #192