fix: prevent unobserved exceptions in bulk export
Sometimes the bulk export in the asset explorer could throw an exception in the asynchronous task, which was unfortunately causing the export to fail silently with no feedback to the user due to an oversight.

This commit changes all the file and folder dialogs in IPlatformService to accept a Task as a callback rather than a void function, so that asynchronous callbacks can be used and properly observed. As such, all calls to these methods have been updated accordingly.