feat: remove weapon guardrails from Steam mods
The binmod builder had a piece of code forcing all vertex weights to be automatically reassigned to C_Body, as for a majority of weapon mods, this is the only viable bone to rig to.

However, this was preventing correct rigging of certain multi-part weapons such as guns and crossbows.

This commit removes this guardrail, as it is more important that users that are using the software correctly are able to make these weapons than it is to prevent those not using it correctly from making mistakes.

Resolves: #204