feat: rework all texture code
Reworks all texture format code, texture conversions, etc. This was necessary due to DirectXTex not being cross-platform, and since it was being reworked anyway, the texture API has been vastly improved.

DirectXTex has been replaced with a combination of ImageSharp for standard non-HDR raster image formats, and NVTT for DDS support and block compression.

Scarlet has been removed from the repository due to implementing a more efficient deswizzling algorithm directly in Flagrum.Core.

Windows versions of NVTT and OpenEXRCore are not present and need to be added before releasing the cross-platform build.

Rudimentary EXR support has been added for HDR textures, but this does not currently have the UI/UX in Flagrum itself to expose this to the user in a meaningful way.